The people that are the hardest to love are usually the ones who need it the most.
