+++ auctions/english-auction/project/contracts/auction-contract/tests/functions/core/create.rs
@@ -1,20 +1,23 @@
 use crate::utils::{
-    abigen_bindings::english_auction_mod::State,
-    asset_abi_calls::mint_and_send_to_address,
-    english_auction_abi_calls::{auction_info, create},
-    nft_abi_calls::{approve, mint, set_approval_for_all},
-    test_helpers::{
-        create_auction_copy, defaults_nft, defaults_token, nft_asset, setup, token_asset,
+    interface::core::{
+        asset::mint_and_send_to_address,
+        auction::create,
+        nft::{approve, mint},
     },
+    setup::{defaults_nft, defaults_token, nft_asset, setup, token_asset},
 };
-use fuels::{
-    prelude::{AssetId, CallParameters, TxParameters},
-    types::Identity,
-};
+use fuels::types::Identity;
 
 mod success {
 
     use super::*;
+    use crate::utils::{
+        interface::{
+            core::nft::{approve, mint, set_approval_for_all},
+            info::auction_info,
+        },
+        setup::{create_auction_copy, State},
+    };
 
     #[tokio::test]
     async fn creates_multiple_auctions() {
@@ -446,6 +449,7 @@ mod success {
 mod revert {
 
     use super::*;
+    use fuels::prelude::{AssetId, CallParameters, TxParameters};
 
     #[tokio::test]
     #[should_panic(expected = "ReserveLessThanInitialPrice")]
