+++ auctions/english-auction/project/contracts/auction-contract/tests/functions/core/mod.rs
@@ -1,0 +1,4 @@
+mod bid;
+mod cancel;
+mod create;
+mod withdraw;
