+++ auctions/english-auction/project/contracts/auction-contract/tests/functions/info/mod.rs
@@ -1,0 +1,3 @@
+mod auction_info;
+mod deposit_balance;
+mod total_auctions;
