+++ auctions/english-auction/project/contracts/auction-contract/tests/functions/mod.rs
@@ -1,7 +1,2 @@
-mod auction_info;
-mod bid;
-mod cancel;
-mod create;
-mod deposit_balance;
-mod total_auctions;
-mod withdraw;
+mod core;
+mod info;
